Added waves and upgrades
difficulty increases every wave and upgrades available every 5 waves

+++ GunBoots/Assets/Scripts/Enemies/EnemyCombat/IEnemyStats.cs
@@ -6,7 +6,7 @@ public interface IEnemyStats
 {
     Vector3 Position { get; }
 
-    float contactDamage { get; }
+    int contactDamage { get; }
 
     void Damage(float damage);
